move func to Crawler
